Change way Capslock is assigned to Ctrl
I couldn't get the Capslock assignment to work as was. So 
instead it makes more sense to me to assign the Capslock
to Fn + Capslock, and then Capslock to Ctrl. To assign Capslock
to Ctrl, moreover, I needed to reference R_Ctrl instead of L_Ctrl,
I think because the L_Ctrl has been set per the order of operations
to Fn already.